Fix types errors
